+++ Core/Library/Interfaces/IEventAggregator.cs
@@ -1,10 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
-using Xamarin.Realm.Service.Events;
 
 namespace Xamarin.Realm.Service.Interfaces
 {
@@ -13,8 +7,6 @@ namespace Xamarin.Realm.Service.Interfaces
     {
         void Raise<TArgs>(string eventName, TArgs eventArgs) where TArgs : EventArgs;
 
-        void AddEvent(string eventName);
-
-        bool RemoveEvent(string eventName);
+        void AddTarget(object target);
     }
 }