+++ Core/Library/Properties/AssemblyInfo.cs
@@ -24,5 +24,5 @@ using System.Reflection;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyInformationalVersion("0.1.5")]
+[assembly: AssemblyInformationalVersion("0.1.6")]
 [assembly: AssemblyFileVersion("0.1.4.0")]